Move image fetching method to separate module.
