+++ src/javascript/controller/weatherControler.js
@@ -1,33 +1,10 @@
-import { createApi } from 'unsplash-js';
 import changeTextColors from '../helpers/changeTextColors';
+import fetchImageData from '../model/imageData';
 
 const { fetchWeatherData } = require('../model/weatherData');
 
 function sleep(ms) {
   return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-const unsplash = createApi({
-  accessKey: 'RmRQetkGdLG3o6Dv0LTcQ9-7Fvi6X1FoTF6Ax_pDsV4',
-  fetch: global.fetch,
-});
-
-// const populateHeader = () => {
-
-// }
-
-async function fetchImageData(queryText) {
-  const result = await unsplash.search.getPhotos({
-    query: queryText,
-    page: 1,
-    perPage: 10,
-  });
-  console.log(result);
-  const index = Math.floor(Math.random() * 10);
-  const img = result.response.results[index].links.download;
-  const avarageColorCode = result.response.results[index].color;
-  // console.log(">2> secondly run this");
-  return [img, avarageColorCode];
 }
 
 async function updateBodyBackground(queryText) {
