--- src/javascript/model/gifData.js
@@ -1,5 +1,0 @@
-const fetchGifData = async () => {
-
-};
-
-export { fetchGifData };
